Moved NavbarInfoButton into its own component
